добавил удаление квартир, preference, обновление БД

+++ src/main/java/com/example/hpsus/kotlintest/datajson/Apartment.kt
@@ -1,3 +1,3 @@
 package com.example.hpsus.kotlintest.datajson
 
-data class Apartment(val indiApartment: Int, val indiHome: Int, val floor: Int, val area: Double)
+data class Apartment(val indiApartment: Int, val indiHome: Int, val floor: Int, val area: Float)
